moved Nav to separate file
