+++ client/src/App.tsx
@@ -1,4 +1,3 @@
-
 //import * as React from "react"
 import React from "react";
 import {
@@ -7,10 +6,9 @@ import {
   Route,
   NavLink
 } from "react-router-dom";
-import styled from '@emotion/styled';
-
 
 import Post from "./pages/Post"
+import Nav from "./components/header/Nav"
 
 const App: React.FC = () => {
   return (
@@ -34,41 +32,4 @@ function Home() {
 }
 
 
-const NavUnlisted = styled.ul`
-  display: flex;
-  a {
-    text-decoration: none;
-  }
-  li {
-    color: black;
-    margin: 0 0.8rem;
-    font-size: 1.3rem;
-    position: relative;
-    list-style: none;
-  }
 
-  .current {
-    li {
-      border-bottom: 2px solid blue;
-    }
-  }
-`;
-
-
-const links = [
-  {name: "Home",path:"/"},
-  {name: "Property",path:"/property"},
-  {name: "Submit New Property",path:"/submit-new-property"},
-];
-
-function Nav() {
-  return (
-    <NavUnlisted>
-      {links.map((link,index) => (
-        <NavLink key={index} to={link.path} exact activeClassName="current">
-          <li>{link.name}</li>
-        </NavLink>
-      ))}
-    </NavUnlisted>
-  );
-}
