+++ client/src/components/header/Nav.js
@@ -1,0 +1,40 @@
+import styled from '@emotion/styled';
+import { NavLink } from 'react-router-dom';
+
+const links = [
+  { name: 'Home', path: '/' },
+  { name: 'Property', path: '/property' },
+  { name: 'Submit New Property', path: '/submit-new-property' },
+];
+
+export default function Nav() {
+  return (
+    <NavUnlisted>
+      {links.map((link, index) => (
+        <NavLink key={index} to={link.path} exact activeClassName="current">
+          <li>{link.name}</li>
+        </NavLink>
+      ))}
+    </NavUnlisted>
+  );
+}
+
+const NavUnlisted = styled.ul`
+  display: flex;
+  a {
+    text-decoration: none;
+  }
+  li {
+    color: black;
+    margin: 0 0.8rem;
+    font-size: 1.3rem;
+    position: relative;
+    list-style: none;
+  }
+
+  .current {
+    li {
+      border-bottom: 2px solid blue;
+    }
+  }
+`;
